Check if the ID is a number even if it's in Eastern Arabic form
Instead of casting to `Int` as a check if the string is an integer, use
`NumberFormatter` to try and convert the string to a number.
Unlike `Int`'s initializer, `NumberFormatter` is powerful enough to
check against all characters outside ASCII that represent numerals.
Making the formatter's locale `"en"` allows us to define the idString constant
for the rest of the code to continue working.

Fixes: https://github.com/alhazmy13/Saudi-ID-Validator/issues/13